Ajustando analisador léxico
